VerificationDelete

+++ src/components/Groups/Tabs/ItemPanel/Item/Item.jsx
@@ -12,6 +12,7 @@ import { updateFieldItem, deleteFieldItem } from '../../action';
 import { startApiFetch, failApiFetch, successApiFetch } from '../action';
 import withFormikField from '../../../../../HOC/widthFormikField';
 import { setOpen } from '../CantDeleteDialog/action';
+import PreventDelete from './PreventDelete';
 
 const services = new Services();
 
@@ -35,6 +36,10 @@ export const Item = ({
 
   const [visible, setVisible] = useState(false);
 
+  const [open, setOpenModal] = useState(false);
+
+  const [selectedValue, setSelectedValue] = useState(false);
+
   const { enqueueSnackbar } = useSnackbar();
 
   const [itemName, setItemName] = useState('');
@@ -47,12 +52,6 @@ export const Item = ({
   const handleChangeItemPUC = (element) => {
     const editableItem = element.target.value;
     setItemPUC(editableItem);
-  };
-
-  const submit = () => {
-    if (formRef.current) {
-      formRef.current.handleSubmit();
-    }
   };
 
   useEffect(() => {
@@ -99,31 +98,44 @@ export const Item = ({
     }
   };
 
-  const handleDeleteField = async () => {
+  const prepareToDelete = () => {
+    setOpenModal(true);
+  };
+
+  const handleDeleteField = async (readyForDelete = false) => {
     if (disabled) {
-      try {
-        const itemList = [...items];
-        const itemsBeforeDelete = [...itemsFiltered];
-        const fieldToDelete = itemsBeforeDelete[index].id;
-        const response = await services.deleteItem({ id: fieldToDelete });
-        const itemsAfterDelete = itemList.filter(
-          (element) => element.id !== fieldToDelete,
-        );
-        onDeleteField(itemsAfterDelete);
-        setDisabled((prevstate) => !prevstate);
-      } catch (error) {
-        if (error.response.data.message === 'itemAssociate') {
-          onSetOpenCantDelete('item');
-          onSuccessApi();
-        } else {
-          onFailApi();
-          enqueueSnackbar(error.response.data.message, { variant: 'error' });
+      if (readyForDelete) {
+        try {
+          const itemList = [...items];
+          const itemsBeforeDelete = [...itemsFiltered];
+          const fieldToDelete = itemsBeforeDelete[index].id;
+          const response = await services.deleteItem({ id: fieldToDelete });
+          const itemsAfterDelete = itemList.filter(
+            (element) => element.id !== fieldToDelete,
+          );
+          onDeleteField(itemsAfterDelete);
+          setDisabled((prevstate) => !prevstate);
+        } catch (error) {
+          if (error.response.data.message === 'itemAssociate') {
+            onSetOpenCantDelete('item');
+            onSuccessApi();
+          } else {
+            onFailApi();
+            enqueueSnackbar(error.response.data.message, { variant: 'error' });
+          }
         }
       }
     } else {
       setDisabled((prevstate) => !prevstate);
     }
   };
+
+  const handleClose = (value) => {
+    setOpenModal(false);
+    setSelectedValue(value);
+    handleDeleteField(value);
+  };
+
   return (
     <>
       <TableCell component="th" scope="row">
@@ -180,7 +192,7 @@ export const Item = ({
                     />
                   </Button>
                   <Button
-                    onClick={handleDeleteField}
+                    onClick={disabled ? prepareToDelete : handleDeleteField}
                     onMouseEnter={() => setVisible((prevState) => !prevState)}
                     onMouseLeave={() => setVisible((prevState) => !prevState)}
                   >
@@ -192,6 +204,11 @@ export const Item = ({
           )}
         </Formik>
       </TableCell>
+      <PreventDelete
+        selectedValue={selectedValue}
+        open={open}
+        onClose={handleClose}
+      />
     </>
   );
 };
@@ -210,10 +227,6 @@ Item.propTypes = {
 };
 
 const mapStateToProps = (state) => {
-  const current =
-    state.groups.groupItemField.itemsFiltered[
-      state.groups.groupItemField.index
-    ];
   return {
     items: state.groups.groupTabs.items,
     itemsFiltered: state.groups.groupItemField.itemsFiltered,
